replace to postgres

+++ sweater/src/main/resources/db/migration/V4__Add_join.sql
@@ -1,5 +1,5 @@
 create table products_buy (
-    bag_id int8 not null  references bag(id),
-    product_id int8 not null references product(id),
+    bag_id bigserial not null  references bag(id),
+    product_id bigserial not null references product(id),
     primary key (bag_id, product_id)
     );
